CI: Build and deploy the control plane on a lab machine (#1390)
* test deploying the control plane on a lab machine

* [screams in pfexec]

* try out the new target

* XXX [omit] temporarily remove the lengthy CI job

* XXX the lab-netdev ramdisk includes /opt/oxide/opte/bin/opteadm now

* XXX [omit] shorten cycle by reusing existing package

* XXX fix tar flags?

* XXX simon says rm -rf

* XXX rebuild

* XXX [omit] work around 14537 (UFS should not allow directories to be unlinked)

* XXX /var/oxide should be tmpfs?

* XXX backout: [omit] shorten cycle by reusing existing package"

* XXX /var/oxide is a file system not a directory

* XXX /zone not created properly?

* XXX [omit] shorten cycle by reusing existing package (again)

* XXX simon says zfs create

* XXX add some notes, dont uninstall at the end

* XXX backout: [omit] shorten cycle by reusing existing package (again)
XXX backout: [omit] temporarily remove the lengthy CI job

* hell

* poke

* XXX rebuild

* XXX nexus external API now on a special IPv4 address?

* XXX nexus external API now on a special IPv4 address? (with feeling!)

* XXX try nexus for longer, also collect debugging on failure

* add nexus log to outputs

* XXX try nexus for longer, also collect debugging on failure (pfexec)

+++ package/src/bin/omicron-package.rs
@@ -339,9 +339,11 @@ fn do_install(
         Ok(())
     })?;
 
-    // Ensure we start from a clean slate - remove all zones & packages.
-    uninstall_all_packages(config);
-    uninstall_all_omicron_zones()?;
+    if env::var("OMICRON_NO_UNINSTALL").is_err() {
+        // Ensure we start from a clean slate - remove all zones & packages.
+        uninstall_all_packages(config);
+        uninstall_all_omicron_zones()?;
+    }
 
     // Extract all global zone services.
     let global_zone_service_names = config
